Add integer interpret and register algorithm.
This algorithm uses vectors and standard C++ functions to register
integers into the program with name and value corresponding in order.
There's another algorithm that will make the program be able to retrieve
the results using the console out function. Which that algorithm is
still being tested and soon will be committed into the GitHub project
when it is completed.