Add CheckForIntAfterMath Algorithm to Variable.h
This algorithm is used for the retrieval of integers stored in the
program's memory. This is so it can display it on the console using the
console out feature. The integers are stored in the int vector (Int)
which has the name stored in the string vector (Name). These are
corresponding and when the name is found in the (ProcessInput) variable.
The algorithm then checks for console out syntax and outputs it as
console out would in a compiled C++ program. This needs some improvement
in the way the algorithm is written. But for now it serves the purpose
as working code.