Add interpreter include list and iostream prototype
This will be used for so that the interpreter can check if you included
the libraries that your function will need to be able to work. This will
be like it would in the compiled version of C++ code. I'm fixing to
implement a switch to allow this feature to be disabled as well. That
being said I also added another algorithm used to register the include
file of iostream. But that was a prototype and because it needs to be
universal to work with all of the C++ include libraries. This algorithm
will need some touch ups to make it that way.