Migrate Project.h into the Project branch of CPPI
The project manager is a add-on to the CPP-Interpreter. It sets up a
environment to see what your program looks like if it was typed normally
into the actual C++ IDE to make ready for compile. With this add-on, you
will also be able to export projects to be passed into the C++ IDE, so
it could be compiled. It also uses the Windows API to start the
calculator app which you most of the time will use in C++.