Add & Update Documentation. Also fix code format.
The while loop that was used for the CheckForIntAfterMath function was not suited in this case. It is now replaced with a for loop which is suited in this case. This is because of how it incremented the iterator at the end of the while loop. The for loop is more neat and works the same in this case of code.